tests/wpt/web-platform-tests/html/canvas/offscreen/conformance-requirements/2d.coordinatespace.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.coordinatespace</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.coordinatespace</h1>
<p class="desc">Coordinate space goes from top-left to bottom-right</p>

<p class="notes">This should not be upside down.
<script>
var t = async_test("Coordinate space goes from top-left to bottom-right");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.fillStyle = '#00f';
ctx.fillRect(0, 0, 100, 50);
ctx.fillStyle = '#0ff';
ctx.fillRect(0, 0, 50, 25);
_assertPixel(offscreenCanvas, 25,12, 0,255,255,255, "25,12", "0,255,255,255");
_assertPixel(offscreenCanvas, 75,12, 0,0,255,255, "75,12", "0,0,255,255");
_assertPixel(offscreenCanvas, 25,37, 0,0,255,255, "25,37", "0,0,255,255");
_assertPixel(offscreenCanvas, 75,37, 0,0,255,255, "75,37", "0,0,255,255");
t.done();

});
</script>
